Fixed mask and added tqdm.auto
The prepare_mask function now has a return_count option to allow it to
return the total amount of elements that will be left unmasked as this
is a common thing to want to know.

TQDM progress bar now uses 'auto' which means it will detect if it is in
a shell of jupyter notebook and display a nice version for that
environment.